requirements.txt
langchain
python-dotenv
langchain-openai
langchain_community
faiss-cpu
streamlit
langchain_groq
arxiv
wikipedia
sentence_transformers
faiss_cpu
pandas
openai
duckdb
langchain_huggingface

langchain 
langchain_core 
langchain_groq
langchain_community 
langchain 
langgraph